+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 # Work In Progress
 
 - Minecraft 26.1+ support.
-- Added a dynamic mouse smoothing feature. (Thanks, `@archiso86`!) ([PR #100](<https://github.com/Mirsario/Minecraft-CameraOverhaul/pull/100>))
+- Added a dynamic global camera smoothing feature. (Thanks, `@archiso86`!) ([PR #100](<https://github.com/Mirsario/Minecraft-CameraOverhaul/pull/100>))
   Fully configurable. On default settings, this will only be active during sprinting, flying, and underwater swimming.
 - Added 'Sprinting' configuration context.
 - Configuration contexts will now be smoothly blended between, preventing sudden snaps upon elytra/swimming/sprinting activation.